Add LICENSE file and call out license in package.
The homepage specifies the license, however, for systems that do
validation on an actual license file (or something that calls it out),
it's nice to have the file included.  Additionally, it's nice when the
license is called out in PyPI, which should be the case with the
setup.py changes.