Remove logging
In a system under moderate load, the logging is a significant source of
load, even if startLogging was never called.  Profiling was performed
using cProfile.

It is possible that in other uses, the overhead is acceptable, so I am
not planning to submit this upstream, however, for my use case, it is
not.

This is, at this point, a total fork from the upstream (for reasons
surrounding the ipaddress dependency addition as I noted in an issue
question).